Setup the init module and changed the reorder to work in a function
